se crea script para generar json y se agrega conteo de notas al script parser
